fix(m2ee/config.py) remove WARNING: Mendix Runtime not found for version
We need to fix this properly at a later time. However, since the M2EE
lib here is currently a copy, this is the most pragmatic way to remove
this warning, which is probably confusing for the users of our buildpack.
The warning is a valid thing (because at that point in time, there is
no Mendix Runtime), however, start.py takes care of ensuring the
Mendix Runtime is actually in place before we actually start the app.