implement bluemix object storage
note that the "container" argument should probably be improved, it's a
user configurable namespace for objects, but there is no way for the
user to configure it here. Right now it we just use the name of the cf
service, but if that includes spaces or other characters this might
break. We could also just hard code it to "mendix" or something.

for this implementation a split between S3 based configuration and Swift
object store configuration needed to be introduced.