allow swift container name to be specified, default to "mendix"
By default Bluemix Object Storage gives you a file store with no
containers. An object needs to be stored in a container, so we need to
pick a container name ourselves. We default to "mendix" but the user can
control this via the SWIFT_CONTAINER_NAME environment variable.

The runtime will auto-create the container for you (this is controllable
with the AutoCreate flag, which we always put to True).