Use hard coded location path for license file
because HOME envvar points to /home/vcap/app while we expect it to
be /home/vcap. Since we assume /home/vcap to be the homedir in a lot of
places already (see for instance m2ee.yaml) this is acceptable.
Alternatively we can make the location configurable or relative (../) to
the start.py script. But that's more error prone. This issue was raised
because the HOME envvar has different value than when run in Docker.