move from suffix to prefix for filedocuments
This is part of a long migration. We can now safely let the runtime
write filedocuments to prefix based objects in the s3 bucket.

The way we do this is hacky, but we need to support previous runtime
versions too. In the future the runtime can make a prefix option and
then we can have two code paths, one with the hacky way of setting the
prefix, one with the proper if runtime version is newer.