start.py: git fetch a specific tag / runtime version
By fetching a specific tag we only transfer the minimal necessary amount of
data to get the runtime/mxbuild version we want. Also, when the version we want
is not present it already fails when fetching, which makes it fallback faster.

Changing this makes it also compatible with a local shallow repository (git
fetch --tags, wouldn't fetch anything in that case).