add test for app crashing on startup
Currently, when app fails to start, we try to exit but because there are
still subprocesses in the container, CF does not detect a failure and
depends on the health check timeout to put the app in CRASHED state.

We can detect this by using a high timeout and seeing "health check
never passed" in the logs.