tests: add start timeout to app
This is only for the cf cli application, we tried this approach to test
the crashing app, but it turned out that the actual setting we needed
was the health_check_timeout. However, we developed setting environment
variables to the command, which can be a useful addition in the future.