if jdbc config is used, pass it directly to the runtime
There are more complex configuration possibilities with jdbc, such as
encryption settings, which we can't capture in the custom runtime
settings. That's why we pass it through directly. This is only used
for for Azure at the moment, and certain features such as schnapps
integration or metrics integration, which are used in Mendix Cloud,
will not work.