set MALLOC_ARENA_MAX to a lower dynamic value to prevent OOM
The default behavior is 8x the number of detected CPUs . As Cloud
Foundry typically uses large host machines with smaller containers,
and the Java process is unaware of the difference in allocated CPUs,
the numbers are way off. This often leads to high native memory usage,
followed by a cgroup OOM killer event.

We go with Heroku's recommendation of lowering to a
setting of 2 for small instances. We also grown the setting linearly
with memory to be more in line with the default setting in Mendix
Cloud v3.

References:
- https://github.com/cloudfoundry/java-buildpack/issues/163
- https://devcenter.heroku.com/articles/testing-cedar-14-memory-use
- https://github.com/cloudfoundry/java-buildpack/issues/320