Set DatabaseUseSsl to true if sslmode value is set.
Official jdbc documentation:
sslmode = String
possible values include "disable", "require", "verify-ca" and
"verify-full", "allow" and "prefer" will throw an exception.
"require" will default to a non validating SSL factory and not
check the validity of the certificates. "verify-ca" and
"verify-full" use a validating SSL factory and will check that
the ca is correct and the host is correct.

Because we do not pass the original jdbc url we have to decompose the
url query params from `extra`. After discussing with Robert-Jan we
decided to support only `require` which translates to `DatabaseUseSsl=true`

source: https://jdbc.postgresql.org/documentation/head/connect.html