CLDOPS-3676: Supress logging in case of failure to collect metrics.
Customers experiences issue with this a lot.
However, it is my assessment that our customers are only complaining
because of the presence of a message
that contains the word error - they do not experience degraded
functionality of their application. Thus, to ensure our customers only
receive "error" messages when something actionable is wrong, I am
surpressing this error message.