Create multi stage Travis build with seprate lint and test
First and foremost we want to make sure we run the tests with the very
same Python version as we run in production. However we do not want to
lose the benefits of using `black` as an autoformatter and verification
tool. In order to have both we need to separate linting and testing
stages and run them on different Python versions (since `black` does not
support to be installed <Python 3.6, but still has the option to run
*against* Python 3.4 code, see the --target-version flag).