Remove .keys() call.
Calling .keys() will construct an iterator over the keys of the entire dictionary. Testing for membership of the dictionary will call the underlying __contains__ method, which (on a dictionary for sure) will definitely be implemented with Clever Magic which most likely avoids creating an object with all the keys, and simply scans the hashmap.