DVAS-1155: Fix guess_java_version function for Java 11.
Add unit tests to cover all of the strange legacy edge cases. Refactor a
little to make the function more unit testable.

Add appropriate comments for runtime_version 6

Use the attribute 'major' of the runtime_version

Instead of doing a floor to parse the major version we could simply
use the attribute of the runtime_version.

Bump version 4.5.5 → 4.5.6