Account for memory pool name variations
Memory pool names vary with the garbage collector enabled by the JVM.
For instance if we consider 'Eden Space' pool name, it shows up as:
- PS Eden Space in Parallel GC
- G1 Eden Space in G1 GC
- PAR Eden Space in CMS GC

So account for these variations in the name and ensure
we use a standard name in our memory pool mapping.

Add unittests to assert this behaviour.