Bug fix to address negative metrics issue
- Code cache is allocated out of native memory and
is not part of the java heap. So as previously thought
we should not remove the code cache space from java heap
when calculating the unused java heap.

- Similarly, since Java 8 metaspace (or permanent gen)
is allocated out of the native memory. So we should not
negate that space from the native memory calculations.

Update comment with link to JVM troubleshooting doc.

Bump version 4.6.1 → 4.6.2