feat(jdbc): set default sslmode to "verify-full" if running on AWS RDS
To satisfy requirements of PCI, a connection to a DB instance must be
encrypted with SSL/TLS.

With this change JDBC connection to a PostgreSQL database instances
running on AWS RDS will be encrypted using "verify-full" option by
default. To identify if we are running on AWS RDS database host
parameter is matched against "*.rds.amazonaws.aws" pattern.

DatabaseJdbcUrl parameter for PostgreSQL databases only supported
by Mendix ≥ 7, so this change will enable SSL/TLS encryption only for
the applications that are running with Mendix ≥ 7.
https://github.com/mendix/docs/blob/d23111e40e34b3aa8d13943745f150e09c0efee0/content/refguide6/custom-settings.md#41-common-settings

Additionally update PostgreSQL JDBC driver to use the custom path for
the SSL root certificate.
https://jdbc.postgresql.org/documentation/head/connect.html

Signed-off-by: Andrei Krasnitski <andrei.krasnitski@mendix.com>