COA-735: Update storage service configuration
This also relates to RTO-3039.

The behaviour of `PerformDeleteFromStorage` changed in 9.6.10 / 9.12.
There are now three options:
* NoFiles
* TemporaryFiles (formerly False)
* AllFiles (formerly True)

`dont_perform_deletes` implies no deletes, which means
that NoFiles is the appropriate option from this point forward.

Since this will be released with a new MTS release
we'll bump that part of the code while we're at it.