Forward all runtime login metrics via its own output plugin
Adding a separate output plugin just for the runtime login metrics as the volume
of these metrics could be quite large and results in metric buffer overflow.

With a separate output plugin, it can have its own telegraf buffer.
Additionally, as these metrics are not used for the trends graphs it is better
to forward them to TFR as a separate payload.